splits out schema generation into separate files for runner and in browser schema generation
