+++ .github/workflows/schema.yml
@@ -25,4 +25,4 @@ jobs:
         run: npm install
 
       - name: Run Schema Generator
-        run: node schema-generation/schema-generator/schema-generator.js schema-generation/test 
+        run: node ./schema-generation/schema-generator/generate-schema.js schema-generation/test
