+++ schema-generation/schema-generator/generate-schema-runner.js
@@ -1,12 +1,15 @@
-const fs = require('fs');
-const path = require('path');
-const { parseDropdownValues, generateSchema } = require('./schema-generator');
-
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 function readFileSync(filePath) {
     return fs.readFileSync(filePath, 'utf8');
 }
 
 async function main() {
+    const { parseDropdownValues, generateSchema } = await import('./schema-generator.js');
+
     const args = process.argv.slice(2);
     if (args.length > 1) {
         console.error("Usage: node generate-schema.js <output_directory>");
@@ -16,12 +19,15 @@ async function main() {
     const [outputDir] = args;
 
     console.info("Reading dropdown values");
-    const dropdownCsv = readFileSync("/schema-generation/schema-generator/dropdown_values.csv");
+    console.log(path.join(__dirname, "dropdown-values.csv"))
+    const dropdownCsv = readFileSync(path.join(__dirname, "dropdown-values.csv"));
+    //const dropdownCsv = readFileSync("./dropdown-values.csv");
     console.info("Parsing dropdown values");
     const dropdownValues = parseDropdownValues(dropdownCsv);
 
     console.info("Reading questions CSV");
-    const questionsCsv = readFileSync("/schema-generation/schema-generator/pvq-questions.csv");
+    const questionsCsv = readFileSync(path.join(__dirname, "pvq-questions.csv"));
+    //const questionsCsv = readFileSync("./pvq-questions.csv");
 
     const APPLICANT_TYPES = ["NATIONAL_SECURITY"]; // Adjust as needed
